+++ src/Procon.Database.Serialization.Test/TestSerializerModify.cs
@@ -49,7 +49,7 @@ namespace Procon.Database.Serialization.Test {
                     Name = "Score"
                 },
                 new NumericValue() {
-                    Integer = 50
+                    Long = 50
                 }
             });
 
@@ -73,7 +73,7 @@ namespace Procon.Database.Serialization.Test {
                     Name = "Score"
                 },
                 new NumericValue() {
-                    Integer = 50
+                    Long = 50
                 }
             })
             .Condition(new Equals() {
@@ -106,7 +106,7 @@ namespace Procon.Database.Serialization.Test {
                     Name = "Score"
                 },
                 new NumericValue() {
-                    Integer = 50
+                    Long = 50
                 }
             })
             .Condition(new Equals() {
@@ -122,7 +122,7 @@ namespace Procon.Database.Serialization.Test {
                     Name = "Rank"
                 },
                 new NumericValue() {
-                    Integer = 10
+                    Long = 10
                 }
             });
